Time: 8 ms (92.00%), Space: 18.6 MB (37.65%) - LeetHub

+++ 0518-coin-change-ii/0518-coin-change-ii.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int fun(int ind, int target, vector<int>& arr, vector<vector<int>>&dp)
+    int fun(int ind, int target, vector<int>& arr, vector<vector<int>> &dp)
     {
         if(ind == 0){
             return (target % arr[0] == 0);
@@ -8,11 +8,11 @@ public:
         if(dp[ind][target] != -1){
             return dp[ind][target];
         }
-        long notTake = fun(ind-1, target, arr, dp);
+        int notTake = fun(ind-1, target, arr, dp);
         int take = 0;
         if(arr[ind] <= target)
         {
-            take = fun(ind, target-arr[ind], arr,dp);
+            take = fun(ind, target-arr[ind], arr, dp);
         }
         return dp[ind][target] = take + notTake;
     }
